# shop-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/items": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "getItemList",
                "description": "Get all Items",
                "operationId": "903a129b448a6d048872ed19f0cd2416",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Item"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "createItem",
                "description": "Create Item",
                "operationId": "5db62a1456c5428981dccc082b52c734",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "tags": [
                    "Item"
                ],
                "summary": "getItemItem",
                "description": "Get Item",
                "operationId": "9eb823d8efd63b3de273c920f7a42a35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Item"
                ],
                "summary": "updateItem",
                "description": "Update Item",
                "operationId": "a7b3ebb3f8ba639c8460970055f64b7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Item"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Item"
                ],
                "summary": "deleteItem",
                "description": "Delete Item",
                "operationId": "aecdd4fc05d1086d67ea34f9fdeeaa8a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Item": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "added_date",
                    "country_of_origin",
                    "status",
                    "rating",
                    "is_approved",
                    "category_id",
                    "owner_id",
                    "image_url",
                    "contact_info"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the item",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "description": {
                        "description": "Description of the item",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "price": {
                        "description": "Price of the item",
                        "type": "number",
                        "format": "number",
                        "readOnly": false,
                        "nullable": false
                    },
                    "added_date": {
                        "description": "Date item was added",
                        "type": "string",
                        "format": "date",
                        "readOnly": false,
                        "nullable": false
                    },
                    "country_of_origin": {
                        "description": "Country where item was made",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "status": {
                        "description": "Status (e.g., Available, Sold)",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "is_approved": {
                        "description": "Approval status",
                        "type": "boolean",
                        "readOnly": false,
                        "nullable": false
                    },
                    "image_url": {
                        "description": "URL to item image",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    },
                    "contact_info": {
                        "description": "Contact information",
                        "type": "string",
                        "readOnly": false,
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}